% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic_color_manual.R
\name{nordic_color_manual}
\alias{nordic_color_manual}
\title{Update color to use nordic noir color palette.}
\usage{
nordic_color_manual()
}
\value{
Updates colors.
}
\description{
Update color to use nordic noir color palette.
}
