% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic_fill_manual.R
\name{nordic_fill_manual}
\alias{nordic_fill_manual}
\title{Update fill to use nordic noir color palette.}
\usage{
nordic_fill_manual()
}
\value{
Updates colors.
}
\description{
Update fill to use nordic noir color palette.
}
